[tests] - merging the rest setup into the fake client - fixing the fake rest api to localhost only - highlighted by a PR (https://github.com/gambol99/go-marathon/pull/13), makes sense to force the / and trim if required on the ID - will clean it up later - I'd forgo using waitondeployments for now
[wait on deployments]
the problem with using deployment id's is;
one) from > 0.8.0 you can be handed a deployment Id on creation, but it may or may not exist in /v2/deployments
two) there is NO WAY of checking if a deployment Id was successful (i.e. no history). So we can poll /deployments and
end up in the situation of; it's not there?, was it sucessfull? has it not been scheduled yet? should i wait for a second
to see if the deployment starts? or have i missed it between the intervals? ... So instead, i need to poll
the application and check the required instances are inline with the running instances (tasks)

For groups a similar requirement is necessary, although we have to take into account running deployments and Marathon appears
to use the instance count as a means of handling the dependencies between apps