Make Application.{Arg,DependsOn}, Docker.Expose* variadic.
The following unary functions can be invoked multiple times to
effectively set a collection of values:

- Application.Arg()
- Application.DependsOn()
- Docker.Expose()
- Docker.ExposeUDP()

This change makes the before-mentioned functions variadic so that
multiple values can be passed more conveniently.

Note that Arg() should actually be called Args() now. This requires a
bit more refactoring, however, as the struct receiver the method is
defined on already holds a field by the same name. In order to keep the
change set manageable, we'll leave the task of renaming the method to a
separate PR.