Make TestUtilsGetInterfaceAddress platform-independent.
`TestUtilsGetInterfaceAddress()` in `utils_test.go` relies on the
availability of a `lo` interface containing `127.0.0.1` as its first
address member. At least on Mac OS X, this does not work because the
localhost device is named `lo0` and first address is the IPv6 one
(`::1`).

The existing test is modified to first determine a valid address and
subsequently runs the test with the expectation set against the found
address. Along the way, we break out and test a helper function needed
by both the implementation and the updated
`TestUtilsGetInterfaceAddress()` test.