Support Marathon 0.15.0+ JSON definition for 422 error responses.
The `errors` poperty key for 422 HTTP error codes was renamed to
`details` in Marathon 0.15.0. (See the diff on Github.com [1].)  In
order to support both old and newer Marathon versions, we duplicate the
details JSON sub-structure in the overall JSON structure under both the
old and new JSON property key.

As a minor cleanup along the way, we remove spaces from a
space/tab-mixed intentation and align properly in related tests.

[1]
https://github.com/mesosphere/marathon/compare/v0.14.2...v0.15.0#diff-96942d2d65f1b0af92336d1786dd6e64L260