Remove extra pointer layers when passing to r.api*.
We occasionally define a pointer to a type and pass the address of that
pointer to r.api*. While json.Marshal is smart enough to unwrap nested
pointers until it hits `nil` or a type to parse, we shouldn't rely on
this and avoid confusing users.

We also prefer new over var declarations for consistency reasons (and
because new is used more dominantly in the code base).