Simplify debug logger to single function. (#328)
Even with no ConfigOutput set, the debug logger is fairly expensive on
many calls or large input parameters due to the fact that the used
log.Logger implementation requires an internal lock. We reduce
contention and optimize performance by wrapping the log package with a
single self-defined logging function we can configure as a no-op when
needed.